+++ docs/wasm-led/pkg/wasm_led_bg.wasm.d.ts
@@ -1,0 +1,11 @@
+/* tslint:disable */
+/* eslint-disable */
+export const memory: WebAssembly.Memory;
+export function __wbg_ledstate_free(a: number): void;
+export function __wbg_get_ledstate_state(a: number): number;
+export function __wbg_set_ledstate_state(a: number, b: number): void;
+export function __wbg_get_ledstate_duration(a: number): number;
+export function __wbg_set_ledstate_duration(a: number, b: number): void;
+export function on_module_load(): void;
+export function get_led_state_and_duration(a: number, b: number): number;
+export function __wbindgen_start(): void;