+++ src/prelude.rs
@@ -5,8 +5,10 @@
 //! use range_set_blaze::prelude::*;
 //! ```
 pub use crate::{
-    intersection_dyn, intersection_map_dyn, union_dyn, union_map_dyn, CheckSortedDisjoint,
+    intersection_dyn, intersection_dyn, intersection_map_dyn, union_dyn, union_dyn, union_map_dyn,
+    AssumeSortedStarts, CheckSortedDisjoint, CheckSortedDisjoint, DynSortedDisjoint,
     DynSortedDisjoint, DynSortedDisjointMap, MultiwayRangeMapBlaze, MultiwayRangeSetBlaze,
-    MultiwayRangeSetBlazeRef, MultiwaySortedDisjoint, MultiwaySortedDisjointMap, RangeMapBlaze,
-    RangeSetBlaze, SortedDisjoint, SortedDisjointMap,
+    MultiwayRangeSetBlaze, MultiwayRangeSetBlazeRef, MultiwayRangeSetBlazeRef,
+    MultiwaySortedDisjoint, MultiwaySortedDisjoint, MultiwaySortedDisjointMap, RangeMapBlaze,
+    RangeSetBlaze, RangeSetBlaze, SortedDisjoint, SortedDisjoint, SortedDisjointMap, SortedStarts,
 };