+++ src/tests.rs
@@ -1089,7 +1089,7 @@ fn lib_coverage_6() {
 
 #[test]
 fn merge_coverage_0() {
-    let a = CheckSortedDisjoint::new(vec![1..=2, 5..=100].into_iter());
+    let a = CheckSortedDisjoint::new(vec![1..=2, 5..=100]);
     let b = CheckSortedDisjoint::from([2..=6]);
     let m = Merge::new(a, b);
     let n = m.clone();
@@ -1099,9 +1099,9 @@ fn merge_coverage_0() {
     assert!(union1.equal(union2));
     assert!(format!("{p:?}").starts_with("Merge"));
 
-    let a = CheckSortedDisjoint::new(vec![1..=2, 5..=100].into_iter());
-    let b = CheckSortedDisjoint::new(vec![2..=6].into_iter());
-    let c = CheckSortedDisjoint::new(vec![-1..=-1].into_iter());
+    let a = CheckSortedDisjoint::new(vec![1..=2, 5..=100]);
+    let b = CheckSortedDisjoint::new(vec![2..=6]);
+    let c = CheckSortedDisjoint::new(vec![-1..=-1]);
     let m = KMerge::new([a, b, c]);
     let n = m.clone();
     let p = n.clone();
@@ -1113,7 +1113,7 @@ fn merge_coverage_0() {
 
 #[test]
 fn not_iter_coverage_0() {
-    let a = CheckSortedDisjoint::new(vec![1..=2, 5..=100].into_iter());
+    let a = CheckSortedDisjoint::new(vec![1..=2, 5..=100]);
     let n = NotIter::new(a);
     let p = n.clone();
     let m = p.clone();
@@ -1170,16 +1170,16 @@ fn sorted_disjoint_coverage_0() {
     let a = CheckSortedDisjoint::<i32, _>::default();
     assert!(a.is_empty());
 
-    let a = CheckSortedDisjoint::new([1..=2, 5..=100].into_iter());
-    let b = CheckSortedDisjoint::new([1..=2, 5..=100].into_iter());
-    assert!((a & b).equal(CheckSortedDisjoint::new([1..=2, 5..=100].into_iter())));
+    let a = CheckSortedDisjoint::new([1..=2, 5..=100]);
+    let b = CheckSortedDisjoint::new([1..=2, 5..=100]);
+    assert!((a & b).equal(CheckSortedDisjoint::new([1..=2, 5..=100])));
 
-    let a = CheckSortedDisjoint::new([1..=2, 5..=100].into_iter());
-    let b = CheckSortedDisjoint::new([1..=2, 5..=100].into_iter());
+    let a = CheckSortedDisjoint::new([1..=2, 5..=100]);
+    let b = CheckSortedDisjoint::new([1..=2, 5..=100]);
     assert!((a - b).is_empty());
 
-    let a = CheckSortedDisjoint::new([1..=2, 5..=100].into_iter());
-    let b = CheckSortedDisjoint::new([1..=2, 5..=100].into_iter());
+    let a = CheckSortedDisjoint::new([1..=2, 5..=100]);
+    let b = CheckSortedDisjoint::new([1..=2, 5..=100]);
     assert!((a ^ b).is_empty());
 }
 
@@ -1211,7 +1211,7 @@ fn sorted_disjoint_coverage_1() {
 #[should_panic]
 fn sorted_disjoint_coverage_2() {
     #[allow(clippy::reversed_empty_ranges)]
-    let mut a = CheckSortedDisjoint::new([1..=0].into_iter());
+    let mut a = CheckSortedDisjoint::new([1..=0]);
     a.next();
 }
 
@@ -1219,7 +1219,7 @@ fn sorted_disjoint_coverage_2() {
 #[should_panic]
 fn sorted_disjoint_coverage_3() {
     #[allow(clippy::reversed_empty_ranges)]
-    let mut a = CheckSortedDisjoint::new([1..=1, 2..=2].into_iter());
+    let mut a = CheckSortedDisjoint::new([1..=1, 2..=2]);
     a.next();
     a.next();
 }
@@ -1228,28 +1228,28 @@ fn sorted_disjoint_coverage_3() {
 #[should_panic]
 fn sorted_disjoint_coverage_4() {
     #[allow(clippy::reversed_empty_ranges)]
-    let mut a = CheckSortedDisjoint::new([0..=i128::MAX].into_iter());
+    let mut a = CheckSortedDisjoint::new([0..=i128::MAX]);
     a.next();
 }
 
 #[test]
 fn sorted_disjoint_iterator_coverage_0() {
-    let a = CheckSortedDisjoint::new([1..=2, 5..=100].into_iter());
-    let b = CheckSortedDisjoint::new([1..=2, 5..=101].into_iter());
+    let a = CheckSortedDisjoint::new([1..=2, 5..=100]);
+    let b = CheckSortedDisjoint::new([1..=2, 5..=101]);
     assert!(b.is_superset(a));
 }
 
 #[test]
 fn union_iter_coverage_0() {
-    let a = CheckSortedDisjoint::new(vec![1..=2, 5..=100].into_iter());
-    let b = CheckSortedDisjoint::new(vec![1..=2, 5..=101].into_iter());
+    let a = CheckSortedDisjoint::new(vec![1..=2, 5..=100]);
+    let b = CheckSortedDisjoint::new(vec![1..=2, 5..=101]);
     let c = a.union(b);
     assert!(format!("{c:?}").starts_with("UnionIter"));
 }
 
 #[test]
 fn unsorted_disjoint_coverage_0() {
-    let a = AssumeSortedStarts::new([1..=2, 5..=100].into_iter());
+    let a = AssumeSortedStarts::new([1..=2, 5..=100]);
     assert!(format!("{a:?}").starts_with("AssumeSortedStarts"));
 }